Improve tests
